requirements.txt
langchain
langchain-community
openai
python-dotenv
streamlit
fpdf2
python-docx
requests